cli arguments
```
 go run main.go -h
NAME:
   rekor-phren is a tool to update the BigQuery table and the bucket with the rekor entries - rekor-phren update -u <rekor url> -b <bucket name> -t <table name> -s <start> -e <end> -c <concurrency>

USAGE:
   rekor-phren is a tool to update the BigQuery table and the bucket with the rekor entries [global options] command [command options] [arguments...]

COMMANDS:
   update   update -u <rekor url> -b <bucket name> -t <table name> -s <start> -e <end> -c <concurrency>
   help, h  Shows a list of commands or help for one command

GLOBAL OPTIONS:
   --bigquery-table-name value, -t value       (default: "rekor_test") [$PHREN_TABLE]
   --concurrency-level-update value, -c value  (default: 10) [$PHREN_CONCURRENCY]
   --end-index value, -e value                 (default: 0) [$PHREN_END]
   --gcs-bucket-name value, -b value           (default: openssf-rekor-test) [$PHREN_BUCKET]
   --help, -h                                  show help (default: false)
   --number-of-retries value, -r value         (default: 5) [$PHREN_RETRY]
   --rekor-url value, -u value                 (default: https://api.rekor.sigstore.dev) [$REKOR_URL]
   --start-index value, -s value               (default: 0) [$PHREN_START]

```

+++ pkg/types.go
@@ -64,14 +64,14 @@ type importHashedrekord struct {
 }
 
 type tlog struct {
-	TreeSize int64 `json:"treeSize"`
+	TreeSize int `json:"treeSize"`
 	host     string
 }
 
 // TLog holds current root hash and size of the merkle tree used to store the log entries.
 type TLog interface {
-	Size() (int64, error)
-	Entry(index int64) (Entry, error)
+	Size() (int, error)
+	Entry(index int) (Entry, error)
 }
 type RekordData struct {
 	Hash RekorDataHash `json:"hash"`
